+++ app/controller/search_params_model.py
@@ -3,14 +3,14 @@ from datetime import date
 
 from pydantic import BaseModel, field_validator, model_validator
 
-from app.request.field_values import (
+from app.controller.field_values import (
     FIELD_LENGTHS,
     NUMERIC_FIELD_LIMITS,
     VALID_ADMIN_FIELDS_TO_SELECT,
     VALID_FIELD_VALUES,
     VALID_FIELDS_TO_SELECT,
 )
-from app.request.helpers import (
+from app.utils.helpers import (
     check_params_are_none_except_excluded,
     clean_str,
     match_bool_to_insee_value,
