+++ app/service/convention_collective.py
@@ -8,7 +8,7 @@ from app.elastic.parsers.siren import is_siren
 from app.elastic.queries.search_by_siren import search_index_by_siren
 from app.exceptions.siren import InvalidSirenError
 from app.models.response import CcResponseModel
-from app.response.formatters.convention_collective import (
+from app.service.formatters.convention_collective import (
     extract_idcc_siret_mapping_from_ul,
 )
 from app.utils.cache import cache_strategy
