+++ app/service/formatters/elus.py
@@ -1,6 +1,5 @@
 from app.models.unite_legale import Elu
-from app.response.helpers import get_value
-from app.utils.helpers import convert_to_year_month
+from app.utils.helpers import convert_to_year_month, get_value
 
 
 def format_elus(elus=None):
